\name{cw_stripped}
\docType{data}
\alias{cw_stripped}
\title{Connecting Words model}
\description{
  This model predicts Connecting Words scores from the rest of the form.
  The underlying data has been stripped from the model to save space.
  Models age.
}
\usage{cw_stripped}
\format{An lm() model suitable for predict().}
\source{Stanford Wordbank.}
\references{

  Frank, M. C., Braginsky, M., Yurovsky, D., & Marchman, V. A. (2017).
  Wordbank: An open repository for developmental vocabulary data.
  \emph{Journal of Child Language},
  44(3), 677–694.
  https://doi.org/10.1017/S0305000916000209

}
\keyword{datasets}