\name{s_dict}
\docType{data}
\alias{s_dict}
\title{Words and Sentences dictionary}
\description{
    This table contains all of the items from Word and Sentences and their
    category.
}
\usage{s_dict}
\format{Data frame}
\source{Stanford Wordbank.}
\references{

  Frank, M. C., Braginsky, M., Yurovsky, D., & Marchman, V. A. (2017).
  Wordbank: An open repository for developmental vocabulary data.
  \emph{Journal of Child Language},
  44(3), 677–694.
  https://doi.org/10.1017/S0305000916000209

}
\keyword{datasets}