% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{total_WG_to_WS_noage_stripped}
\alias{total_WG_to_WS_noage_stripped}
\title{Predict total score}
\format{
A linear model

A linear model predicting WS score from WG score (no age). The embedded
data have been stripped from the object.
}
\usage{
total_WG_to_WS_noage_stripped
}
\description{
Predict total score
}
\references{
Frank, M. C., Braginsky, M., Yurovsky, D., & Marchman, V. A. (2017).
Wordbank: An open repository for developmental vocabulary data.
\emph{Journal of Child Language}, 44(3), 677-694.
\doi{10.1017/S0305000916000209}
}
\keyword{datasets}
