% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wg2ws_items.R
\name{wg2ws_items}
\alias{wg2ws_items}
\title{Title}
\usage{
wg2ws_items(items, error_on_missing = TRUE, in_inside = "either")
}
\arguments{
\item{items}{List of WG items present for individual}

\item{error_on_missing}{If TRUE, check whether all items are actual WG items}

\item{in_inside}{"In" and "inside" appear as two items on WG, but one
("inside/in") on WS. If "either," treat "inside/in" as endorsed if either
appears. For "in" or "inside", treat "inside/in" as endorsed based on
the presence of the indicated item.}
}
\value{
A data frame with 22 rows indicating item totals for all WS
categories.
}
\description{
Title
}
\examples{

items <- c("woof woof", "candy", "cry", "sheep", "pattycake", "lamp", "hi")
wg2ws_items(items)

       category n
1 action_words 3
2      animals 0
3   body_parts 0
             ...
}
