% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wg2ws_list_items.R
\name{wg2ws_list_items}
\alias{wg2ws_list_items}
\title{List instrument items}
\usage{
wg2ws_list_items(instrument)
}
\arguments{
\item{instrument}{"WG" or "WS"}
}
\value{
List of items
}
\description{
List instrument items
}
\details{
Simply list the items from each instrument for convenience.
}
\examples{
wg2ws_list_items("WG")
wg2ws_list_items("WS")
}
