% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wg2ws_summarize_cat.R
\name{wg2ws_summarize_cat}
\alias{wg2ws_summarize_cat}
\title{Summarize category table}
\usage{
wg2ws_summarize_cat(category_scores)
}
\arguments{
\item{category_scores}{A 22x2 category result table}
}
\value{
A three column data frame, with total score, lexical, and syntactic
scores
}
\description{
Summarize category table
}
