% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wg2ws_total_age.R
\name{wg2ws_total_age}
\alias{wg2ws_total_age}
\title{wg2ws_total_age}
\usage{
wg2ws_total_age(WG, age = NA)
}
\arguments{
\item{WG}{Words and Gestures total score.}

\item{age}{Age in months (optional). A different, more accurate model is
used if age is supplied.}
}
\value{
Adjusted score, rounded to the nearest integer. Does not return
values below 0 or greater than 680.
}
\description{
wg2ws_total_age
}
