% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat_models_stripped}
\alias{cat_models_stripped}
\title{Predict category scores}
\format{
An object of many linear models.

Linear models predicting WS category scores from WG score and age. The
embedded data have been stripped from the object.
}
\usage{
cat_models_stripped
}
\description{
Predict category scores
}
\references{
Frank, M. C., Braginsky, M., Yurovsky, D., & Marchman, V. A. (2017).
Wordbank: An open repository for developmental vocabulary data.
\emph{Journal of Child Language}, 44(3), 677-694.
\doi{10.1017/S0305000916000209}
}
\keyword{datasets}
