% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_noage_stripped}
\alias{cw_noage_stripped}
\title{Predict Connecting Words scores (no age)}
\format{
Linear model

A linear model predicting Connecting Words scores from other scores
(no age). The embedded data have been stripped from the object.
}
\usage{
cw_noage_stripped
}
\description{
Predict Connecting Words scores (no age)
}
\references{
Frank, M. C., Braginsky, M., Yurovsky, D., & Marchman, V. A. (2017).
Wordbank: An open repository for developmental vocabulary data.
\emph{Journal of Child Language}, 44(3), 677-694.
\doi{10.1017/S0305000916000209}
}
\keyword{datasets}
