% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_stripped}
\alias{cw_stripped}
\title{Predict Connecting Words scores}
\format{
Linear model

A linear model predicting Connecting Words scores from other scores and
age. The embedded data have been stripped from the objects.
}
\usage{
cw_stripped
}
\description{
Predict Connecting Words scores
}
\references{
Frank, M. C., Braginsky, M., Yurovsky, D., & Marchman, V. A. (2017).
Wordbank: An open repository for developmental vocabulary data.
\emph{Journal of Child Language}, 44(3), 677-694.
\doi{10.1017/S0305000916000209}
}
\keyword{datasets}
