% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_stripped}
\alias{cw_stripped}
\title{Connecting Words model}
\description{
Model to predict Connecting Words scores from other scores, given an age.
The data has been stripped from the model.
}
\references{
Frank, M. C., Braginsky, M., Yurovsky, D., & Marchman, V. A. (2017).
Wordbank: An open repository for developmental vocabulary data.
\emph{Journal of Child Language}, 44(3), 677–694.
https://doi.org/10.1017/S0305000916000209
}
\keyword{model}
