% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_dict}
\alias{s_dict}
\title{WS dictionary: items, categories}
\format{
A data frame with 680 rows and 4 columns:

\describe{

\item{category}{Name of category}
\item{item_definition}{Item label, e.g. "baa baa"}
\item{item_id}{Unique ID}
\item{item_kind}{Type of item, only "word"}

}
}
\usage{
s_dict
}
\description{
WS dictionary: items, categories
}
\references{
Frank, M. C., Braginsky, M., Yurovsky, D., & Marchman, V. A. (2017).
Wordbank: An open repository for developmental vocabulary data.
\emph{Journal of Child Language}, 44(3), 677-694.
\doi{10.1017/S0305000916000209}
}
\keyword{datasets}
